Huge Commit

+++ Backend/FileshareBackend/Services/FileSystemService.cs
@@ -1,7 +1,9 @@
+using System.Text;
+using System.Text.Json;
+using DAL.Entities;
 using FileshareBackend.Exceptions;
 using FileshareBackend.Models;
 using FileshareBackend.Services.Interfaces;
-using MimeTypes;
 
 namespace FileshareBackend.Services;
 
@@ -12,10 +14,11 @@ public class FileSystemService : IFileSystemService
     {
         _config = config;
     }
-    public FSEntryModel GetFromDirectory(string[] pathArr)
+    public FSEntryModel GetFromDirectory(string[] pathArr, User? user)
     {
         for (int i = 0; i < pathArr.Length; i++)
         {
+            
             pathArr[i] = Utils.CleanString(pathArr[i]);
         }
 
@@ -52,7 +55,67 @@ public class FileSystemService : IFileSystemService
         */
         if (!Directory.Exists(fullPath)) throw new FileSystemException($"{path} not Found");
         string[] directories = Directory.GetDirectories(fullPath).OrderBy(f => f).ToArray();
-        string[] files = Directory.GetFiles(fullPath).OrderBy(f => f).ToArray();
+        string[] files = Directory.GetFiles(fullPath).Where(x => !x.EndsWith("fsmeta.json")).OrderBy(f => f).ToArray();
+        
+        
+        List<string> allList = new List<string>();
+        FSEntryModel entries = new FSEntryModel();
+        MetaFileModel meta = null!;
+        try
+        {
+            meta = GetMeta(fullPath);
+        }
+        catch (FileSystemException e)
+        {
+            throw e;
+        }
+
+        entries.CanEdit = user != null && meta.Owner == user.Username;
+        
+        if (meta.Visibility == "private" && (user != null && user.Username != meta.Owner))
+            throw new FileSystemException("Unauthorized");
+        for (int i = 0; i < directories.Length; i++)
+        {
+            string[] folders = directories[i].Split('/');
+            string name = folders[folders.Length - 1];
+            var dirMeta = meta.Directories.FirstOrDefault(x => x.Name == name);
+            if((dirMeta.Visibility == "private" || dirMeta.Visibility == "unlisted") && (user != null && user.Username != dirMeta.Owner)) continue;
+            DirectoryModel directory = new DirectoryModel()
+            {
+                Name = name,
+                ItemCount = GetItemCount(directories[i])
+            };
+            entries.Directories.Add(directory);
+        }
+
+        for (int i = 0; i < files.Length; i++)
+        {
+            string[] split = files[i].Split('/');
+            string name = split[split.Length - 1];
+            var fileMeta = meta.Files.First(x => x.Name == name);
+            if ((fileMeta.Visibility == "private" || fileMeta.Visibility == "unlisted") &&
+                (user == null || user.Username != fileMeta.Owner))
+            {
+                continue;
+            }
+            FileModel file = new FileModel()
+            {
+                Name = name
+            };
+            
+            entries.Files.Add(file);
+        }
+
+        return entries;
+
+    }
+    
+    //Only used in Utils
+    public FSEntryModel GetFromDirectory(string path)
+    {
+        
+        string[] directories = Directory.GetDirectories(path).OrderBy(f => f).ToArray();
+        string[] files = Directory.GetFiles(path).OrderBy(f => f).ToArray();
         List<string> allList = new List<string>();
         FSEntryModel entries = new FSEntryModel();
         
@@ -84,12 +147,14 @@ public class FileSystemService : IFileSystemService
 
     }
 
-    public void MoveFile(string[] pathArr, string[] newPathArr, string name, bool overwrite)
+    public void MoveFile(string[] pathArr, string[] newPathArr, string name, User user, bool overwrite)
     {
+        if (user == null) throw new FileSystemException("Unauthorized");
         for (int i = 0; i < pathArr.Length; i++)
         {
             pathArr[i] = Utils.CleanString(pathArr[i]);
         }
+        if (!OwnsItem(pathArr == null ? new string[] {name} : pathArr.Append(name).ToArray(), user.Username, true)) throw new FileSystemException("Unauthorized");
 
         string path = "";
         try
@@ -116,19 +181,23 @@ public class FileSystemService : IFileSystemService
         }
 
         newPath = Path.Join(_config["DirectoryRootPath"], newPath, name);
+        var meta = GetMeta(Path.Join(_config["DirectoryRootPath"], path));
         path = Path.Join(_config["DirectoryRootPath"], path, name);
         if (!overwrite && File.Exists(newPath)) throw new FileSystemException("File exists in new location");
         File.Move(path, newPath, overwrite);
-        
+        RemoveMeta(pathArr.Append(name).ToArray(), true);
+        AddMeta(newPathArr.Append(name).ToArray(), true, user.Username);
     }
 
-    public void MoveDirectory(string[] pathArr, string[] newPathArr)
+    public void MoveDirectory(string[] pathArr, string[] newPathArr, User user)
     {
         for (int i = 0; i < pathArr.Length; i++)
         {
             pathArr[i] = Utils.CleanString(pathArr[i]);
         }
 
+        if (!OwnsItem(pathArr, user.Username, false)) throw new FileSystemException("Unauthorized");
+        
         string path = "";
         try
         {
@@ -153,11 +222,14 @@ public class FileSystemService : IFileSystemService
             throw new FileSystemException("Failed to locate directory");
         }
 
+        
         newPath = Path.Join(_config["DirectoryRootPath"], newPath);
         path = Path.Join(_config["DirectoryRootPath"], path);
         try
         {
             Directory.Move(path, newPath);
+            RemoveMeta(pathArr, false);
+            AddMeta(newPathArr, false, user.Username);
         }
         catch (IOException e)
         {
@@ -166,13 +238,15 @@ public class FileSystemService : IFileSystemService
         
     }
 
-    public void RenameItem(string[] pathArr, string newName)
+    public void RenameItem(string[] pathArr, string newName, User user)
     {
         for (int i = 0; i < pathArr.Length; i++)
         {
             pathArr[i] = Utils.CleanString(pathArr[i]);
         }
-
+        if (pathArr[^1] == "fsmeta.json") throw new FileSystemException("Unauthorized");
+        if (newName == "fsmeta.json") throw new FileSystemException("Unauthorized");
+        
         string oldPath = "";
         try
         {
@@ -184,6 +258,7 @@ public class FileSystemService : IFileSystemService
         }
 
         string newPath = "";
+        string[] oldPR = (string[])pathArr.Clone();
         try
         {
             pathArr[pathArr.Length - 1] = newName;
@@ -196,9 +271,16 @@ public class FileSystemService : IFileSystemService
 
         oldPath = Path.Join(_config["DirectoryRootPath"], oldPath);
         newPath = Path.Join(_config["DirectoryRootPath"], newPath);
+        if (Directory.Exists(oldPath) && !OwnsItem(pathArr, user.Username, false))
+            throw new FileSystemException("Unauthorized");
+        if (File.Exists(oldPath) && !OwnsItem(pathArr, user.Username, true))
+            throw new FileSystemException("Unauthorized");
+        bool isFile = !Directory.Exists(oldPath);
         try
         {
             Directory.Move(oldPath, newPath);
+            RemoveMeta(oldPR, isFile);
+            AddMeta(pathArr, isFile, user.Username);
         }
         catch (IOException e)
         {
@@ -208,12 +290,16 @@ public class FileSystemService : IFileSystemService
 
     }
 
-    public void DeleteItem(string[] pathArr)
+    
+
+    public void DeleteItem(string[] pathArr, User user)
     {
         for (int i = 0; i < pathArr.Length; i++)
         {
             pathArr[i] = Utils.CleanString(pathArr[i]);
         }
+
+        if (pathArr[^1] == "fsmeta.json") throw new FileSystemException("Unauthorized");
         string path = "";
         try
         {
@@ -224,16 +310,20 @@ public class FileSystemService : IFileSystemService
         {
             throw new FileSystemException("Failed to locate item");
         }
-
+        
         if (Directory.Exists(path))
         {
-            if (Directory.GetFileSystemEntries(path).Length != 0) throw new FileSystemException("Directory not empty");
-            
+            if (!OwnsItem(pathArr, user.Username, false)) throw new FileSystemException("Unauthorized");
+            if (Directory.GetFileSystemEntries(path).Length > 1 || !Directory.GetFileSystemEntries(path)[0].EndsWith("fsmeta.json")) throw new FileSystemException("Directory not empty");
+            File.Delete(Directory.GetFileSystemEntries(path)[0]);
             Directory.Delete(path);
+            RemoveMeta(pathArr, false);
         } 
         else if (File.Exists(path))
         {
+            if (!OwnsItem(pathArr, user.Username, true)) throw new FileSystemException("Unauthorized");
             File.Delete(path);
+            RemoveMeta(pathArr, true);
         }
         else
         {
@@ -241,12 +331,13 @@ public class FileSystemService : IFileSystemService
         }
     }
 
-    public void UploadFile(string[] pathArr, IFormFile file)
+    public void UploadFile(string[] pathArr, IFormFile file, User user)
     {
         for (int i = 0; i < pathArr.Length; i++)
         {
             pathArr[i] = Utils.CleanString(pathArr[i]);
         }
+        if (file.FileName == "fsmeta.json") throw new FileSystemException("Unauthorized");
         string path = "";
         try
         {
@@ -259,14 +350,103 @@ public class FileSystemService : IFileSystemService
         }
 
         if (File.Exists(Path.Join(path, file.FileName))) throw new FileSystemException("File already exists");
-
+        AddMeta(Path.Join(path, file.FileName).Split('/'), true, user.Username);
         var stream = File.OpenWrite(Path.Join(path, file.FileName));
         
         file.CopyTo(stream);
         
         stream.Close();
     }
-    
+
+    public void ChangeItemVisibility(string[] pathArr, string visibility)
+    {
+        
+    }
+
+    public bool CanDownload(string[] pathArr, User? user)
+    {
+        List<string> t = pathArr.ToList();
+        t.RemoveAt(t.Count - 1);
+        string[] newArr = t.ToArray();
+        var meta = GetMeta(Path.Join(_config["DirectoryRootPath"], string.Join('/', newArr)));
+        var fileMeta = meta.Files.First(x => x.Name == pathArr[^1]);
+        if (fileMeta.Owner == user?.Username || fileMeta.Visibility != "private") return true;
+        return false;
+    }
+
+    private bool OwnsItem(string[] pathArr, string Owner, bool isFile)
+    {
+        string[] parentPath = pathArr.Take(pathArr.Length - 1).ToArray();
+        
+        var meta = GetMeta(Path.Join(_config["DirectoryRootPath"], string.Join('/', parentPath)));
+        if (isFile)
+        {
+            return meta.Files.First(x => x.Name == pathArr[^1]).Owner == Owner;
+        }
+        return meta.Directories.First(x => x.Name == pathArr[^1]).Owner == Owner;
+    }
+
+    public MetaFileModel GetMeta(string path)
+    {
+        var rawText = File.ReadAllText(Path.Join(path,"fsmeta.json"));
+        try
+        {
+            return JsonSerializer.Deserialize<MetaFileModel>(rawText);
+        }
+        catch (Exception e)
+        {
+            throw new FileSystemException("Server-Side file corrupted");
+        }
+        
+    }
+
+    private void AddMeta(string[] itemPath, bool isFile, string Owner)
+    {
+        string[] parentPath = itemPath.Take(itemPath.Length - 1).ToArray();
+        var meta = GetMeta(Path.Join(_config["DirectoryRootPath"], string.Join('/', parentPath)));
+        
+        var fileMeta = new ItemMeta()
+        {
+            Name = itemPath[^1],
+            Owner = Owner,
+            Visibility = "public"
+        };
+        if (isFile)
+        {
+            meta.Files = meta.Files.Append(fileMeta).ToArray();
+        }
+        else
+        {
+            meta.Directories = meta.Directories.Append(fileMeta).ToArray();
+        }
+
+        File.WriteAllText(Path.Join(_config["DirectoryRootPath"], string.Join('/', parentPath), "fsmeta.json"),
+            string.Empty);
+        var stream = File.OpenWrite(Path.Join(_config["DirectoryRootPath"], string.Join('/', parentPath), "fsmeta.json"));
+        stream.Write(Encoding.UTF8.GetBytes(JsonSerializer.Serialize(meta, new JsonSerializerOptions(){WriteIndented = true})));
+        stream.Close();
+        stream.Dispose();
+    }
+
+    private void RemoveMeta(string[] itemPath, bool isFile)
+    {
+        string[] parentPath = itemPath.Take(itemPath.Length - 1).ToArray();
+        var meta = GetMeta(Path.Join(_config["DirectoryRootPath"], string.Join('/', parentPath)));
+        if (isFile)
+        {
+            meta.Files = meta.Files.Where(x => x.Name != itemPath[^1]).ToArray();
+        }
+        else
+        {
+            meta.Directories = meta.Directories.Where(x => x.Name != itemPath[^1]).ToArray();
+        }
+        File.WriteAllText(Path.Join(_config["DirectoryRootPath"], string.Join('/', parentPath), "fsmeta.json"),
+            string.Empty);
+        var stream = File.OpenWrite(Path.Join(_config["DirectoryRootPath"], string.Join('/', parentPath), "fsmeta.json"));
+        stream.Write(Encoding.UTF8.GetBytes(JsonSerializer.Serialize(meta, new JsonSerializerOptions(){WriteIndented = true})));
+        stream.Close();
+        stream.Dispose();
+    }
     private int GetItemCount(string path)
     {
         return Directory.GetFileSystemEntries(path).Length;
